+++ ChoGGi's Library/Locales/StringsPreCSV.lua
@@ -3,8 +3,6 @@ do !Skipped Line! [[BLANKS]]
 !Skipped Line! <color ChoGGi_green>%s</color>
 !Skipped Line! <color ChoGGi_yellow>%s</color>
 "ID", "Text", "Translation"
-302535920001737, ""
-302535920001738, ""
 302535920001739, ""
 302535920001740, ""
 302535920001741, ""
@@ -1672,6 +1670,8 @@ Takes less then a second on a map with 3616 buildings and 54 drone hubs."
 302535920000622, "If you have any buildings that aren't aligned to the hex grids use this."
 end end end end end end end !Skipped Line!
 do do !Skipped Line! [[GameMenu.lua" "GameFunc.lua]] !Skipped Line!
+302535920001737, "Toggle Use All Loading Screens"
+302535920001738, "Some DLC replaces loading screens with their own, enable to use all of them."
 302535920001679, "Tool Tips"
 302535920001665, "Toggle Infopanel Toolbar Constrain"
 302535920001666, "Limits size of infopanel toolbar buttons for those that have too many buttons (and they go off panel)."