+++ Mods ChoGGi/Automatic Stirling Generators/metadata.lua
@@ -1,0 +1,23 @@
+return PlaceObj("ModDef", {
+	"title", "Automatic Stirling Generators",
+	"id", "ChoGGi_AutomaticStirlingGenerators",
+	"steam_id", "3398216734",
+	"pops_any_uuid", "3b9813c2-91c5-4a89-a565-df09434bf1ea",
+	"lua_revision", 1007000, -- Picard
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagGameplay", true,
+	"description", [[
+Automatically open and close your Stirlings. See mod options to pick shifts.
+
+
+Requested by Selpas_98.
+]],
+})