+++ Mods ChoGGi/Change Skin Wasp Drones/Code/Script.lua
@@ -5,15 +5,21 @@ if not g_AvailableDlc.gagarin then
 	return
 end
 
-local skins = {}
-local c = 0
-local EntityData = EntityData
-for key in pairs(EntityData) do
-	if key:find("DroneJapanFlying") then
-		c = c + 1
-		skins[c] = key
-	end
-end
+--~ local skins = {}
+--~ local c = 0
+--~ local EntityData = EntityData
+--~ for key in pairs(EntityData) do
+--~ 	if key:find("DroneJapanFlying") then
+--~ 		c = c + 1
+--~ 		skins[c] = key
+--~ 	end
+--~ end
+
+local skins = {
+	"DroneJapanFlying",
+	"DroneJapanFlying_02",
+	"DroneJapanFlying_03",
+}
 
 function FlyingDrone:GetSkins()
 	return skins