+++ Mods ChoGGi/Change Skin Wasp Drones/metadata.lua
@@ -4,9 +4,9 @@ return PlaceObj("ModDef", {
 	"steam_id", "1653474003",
 	"pops_any_uuid", "7f5565f7-ea07-45b1-b438-ac6be3f3617c",
 	"lua_revision", 1007000, -- Picard
-	"version", 1,
+	"version", 2,
 	"version_major", 0,
-	"version_minor", 1,
+	"version_minor", 2,
 	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {