+++ Mods ChoGGi/Example Loading Screen/metadata.lua
@@ -1,8 +1,6 @@
 return PlaceObj("ModDef", {
 	"title", "Loading Screen Example",
 	"id", "ChoGGi_LoadingScreenExample",
-	"steam_id", "1535171613",
-	"pops_any_uuid", "fe6ade88-cd32-478a-be30-2241dcafb9a5",
 	"lua_revision", 1007000, -- Picard
 	"version", 1,
 	"version_major", 0,
@@ -15,9 +13,6 @@ return PlaceObj("ModDef", {
 	"description", [[
 Adds a single loading screen as an example to make your own pack.
 This will block the old ones, see Script.lua to add them back as well.
-
-You don't need to sub; this is also available at:
-https://github.com/ChoGGi/SurvivingMars_CheatMods/tree/master/Mods%20ChoGGi/Example%20Loading%20Screen
 
 Requested by... A friend.
 ]],